@@ -216,7 +216,7 @@ public:
     if (halt) {
       galois::runtime::evilPhase += 2; // one for reduce and one for broadcast
       if (galois::runtime::evilPhase >=
-          std::numeric_limits<int16_t>::max()) { // limit defined by MPI or LCI
+          static_cast<uint32_t>(std::numeric_limits<int16_t>::max())) { // limit defined by MPI or LCI
         galois::runtime::evilPhase = 1;
       }
     }
