@@ -26,7 +26,7 @@
 #ifndef GALOIS_DIST_GALOIS_H
 #define GALOIS_DIST_GALOIS_H
 
-#include "galois/runtime/Init.h"
+#include "galois/runtime/SharedMem.h"
 #include "galois/runtime/DistStats.h"
 
 #include <string>
@@ -38,11 +38,17 @@ namespace galois {
  * Explicit class to initialize the Galois Runtime.
  * The runtime is destroyed when this object is destroyed
  */
-class DistMemSys : public runtime::SharedMemRuntime<runtime::DistStatManager> {
+class DistMemSys : public runtime::SharedMem<runtime::DistStatManager> {
 public:
-  explicit DistMemSys(void);
+  explicit DistMemSys();
 
-  ~DistMemSys(void);
+  ~DistMemSys();
+
+  DistMemSys(const DistMemSys&) = delete;
+  DistMemSys& operator=(const DistMemSys&) = delete;
+
+  DistMemSys(DistMemSys&&) = delete;
+  DistMemSys& operator=(DistMemSys&&) = delete;
 };
 
 } // namespace galois
