@@ -322,9 +322,9 @@ public:
   typedef size_t size_type;
   typedef ptrdiff_t difference_type;
   typedef value_type reference;
-  typedef const value_type const_reference;
+  typedef value_type const_reference;
   typedef value_type* pointer;
-  typedef const value_type* const_pointer;
+  typedef value_type* const_pointer;
   typedef pointer iterator;
   typedef const_pointer const_iterator;
   const static bool has_value = false;
