@@ -129,29 +129,15 @@ public:
   bool empty() const { return size_ == 0; }
 
   void reserve(size_t n) {
-    // galois::StatTimer StatTimer_reserve("reserve", "PR_BC");
-    // StatTimer_reserve.start();
     if (n > capacity_) {
-      // size_t old_cap = capacity_;
-      if (capacity_ == 0) capacity_ = 1;
-      // else galois::gDebug("PODResizeableArray enlarged");
-      while (capacity_ < n) capacity_ <<= 1;
-      // if(_Realloc)
-        data_ = static_cast<_Tp*>(realloc(data_, capacity_ * sizeof(_Tp)));
-      // else{
-      //   if (data_ == NULL) {
-      //     data_ = alloc_.allocate(capacity_);
-      //   }
-      //   else {
-      //     pointer old_mem = data_;
-      //     data_ = alloc_.allocate(capacity_);
-      //     galois::gDebug("memcpy: ", old_mem, " -> ", data_);
-      //     memcpy(data_, old_mem, old_cap * sizeof(_Tp));
-      //     alloc_.deallocate(old_mem, old_cap);
-      //   }
-      // }
+      if (capacity_ == 0) {
+        capacity_ = 1;
+      }
+      while (capacity_ < n) {
+        capacity_ <<= 1;
+      }
+      data_ = static_cast<_Tp*>(realloc(reinterpret_cast<void*>(data_), capacity_ * sizeof(_Tp)));
     }
-    // StatTimer_reserve.stop();
   }
 
   void resize(size_t n) {
@@ -162,15 +148,6 @@ public:
   void clear() {
     size_ = 0;
   }
-
-  // void release() {
-  //   if (data_ != NULL) {
-  //     alloc_.deallocate(data_, capacity_);
-  //     data_ = NULL;
-  //     capacity_ = 0;
-  //     size_ = 0;
-  //   }
-  // }
 
   // element access:
   reference operator[](size_type __n) { return data_[__n]; }
@@ -189,7 +166,7 @@ public:
   void assign(iterator first, iterator last) {
     size_t n = last - first;
     resize(n);
-    memcpy(data_, first, n * sizeof(_Tp));
+    memcpy(reinterpret_cast<void*>(data_), first, n * sizeof(_Tp));
   }
 
   reference front() { return data_[0]; }
