@@ -32,6 +32,9 @@
 #ifdef __GALOIS_HET_CUDA__
 #include "bfs_push_cuda.h"
 struct CUDA_Context* cuda_ctx;
+#else
+enum { CPU, GPU_CUDA };
+int personality = CPU;
 #endif
 
 constexpr static const char* const regionname = "BFS";
@@ -103,15 +106,17 @@ struct InitializeGraph {
   void static go(Graph& _graph) {
     const auto& allNodes = _graph.allNodesRange();
 
-#ifdef __GALOIS_HET_CUDA__
     if (personality == GPU_CUDA) {
+#ifdef __GALOIS_HET_CUDA__
       std::string impl_str(syncSubstrate->get_run_identifier("InitializeGraph_"));
       galois::StatTimer StatTimer_cuda(impl_str.c_str(), regionname);
       StatTimer_cuda.start();
       InitializeGraph_allNodes_cuda(infinity, src_node, cuda_ctx);
       StatTimer_cuda.stop();
-    } else if (personality == CPU)
+#else
+        abort();
 #endif
+    } else if (personality == CPU)
     {
       galois::do_all(galois::iterate(allNodes.begin(), allNodes.end()),
                      InitializeGraph{src_node, infinity, &_graph},
@@ -144,15 +149,17 @@ struct FirstItr_BFS {
       __end   = 0;
     }
     syncSubstrate->set_num_round(0);
-#ifdef __GALOIS_HET_CUDA__
     if (personality == GPU_CUDA) {
+#ifdef __GALOIS_HET_CUDA__
       std::string impl_str(syncSubstrate->get_run_identifier("BFS"));
       galois::StatTimer StatTimer_cuda(impl_str.c_str(), regionname);
       StatTimer_cuda.start();
       FirstItr_BFS_cuda(__begin, __end, cuda_ctx);
       StatTimer_cuda.stop();
-    } else if (personality == CPU)
+#else
+        abort();
 #endif
+    } else if (personality == CPU)
     {
       // one node
       galois::do_all(
@@ -221,9 +228,8 @@ struct BFS {
       syncSubstrate->set_num_round(_num_iterations);
       dga.reset();
       work_edges.reset();
-#ifdef __GALOIS_HET_CUDA__
       if (personality == GPU_CUDA) {
-
+#ifdef __GALOIS_HET_CUDA__
         std::string impl_str(syncSubstrate->get_run_identifier("BFS"));
         galois::StatTimer StatTimer_cuda(impl_str.c_str(), regionname);
         StatTimer_cuda.start();
@@ -233,9 +239,10 @@ struct BFS {
         dga += __retval;
         work_edges += __retval2;
         StatTimer_cuda.stop();
-      } else if (personality == CPU)
+#else
+        abort();
 #endif
-      {
+      } else if (personality == CPU) {
         galois::do_all(
             galois::iterate(nodesWithEdges), BFS(priority, &_graph, dga, work_edges), galois::steal(),
             galois::no_stats(),
@@ -305,16 +312,17 @@ struct BFSSanityCheck {
     dgas.reset();
     dgm.reset();
 
-#ifdef __GALOIS_HET_CUDA__
     if (personality == GPU_CUDA) {
+#ifdef __GALOIS_HET_CUDA__
       uint64_t sum;
       uint32_t max;
       BFSSanityCheck_masterNodes_cuda(sum, max, infinity, cuda_ctx);
       dgas += sum;
       dgm.update(max);
-    } else
+#else
+        abort();
 #endif
-    {
+    } else {
       galois::do_all(galois::iterate(_graph.masterNodesRange().begin(),
                                      _graph.masterNodesRange().end()),
                      BFSSanityCheck(infinity, &_graph, dgas, dgm),
@@ -404,12 +412,15 @@ int main(int argc, char** argv) {
     BFSSanityCheck::go(*hg, DGAccumulator_sum, m);
 
     if ((run + 1) != numRuns) {
-#ifdef __GALOIS_HET_CUDA__
       if (personality == GPU_CUDA) {
+#ifdef __GALOIS_HET_CUDA__
         bitset_dist_current_reset_cuda(cuda_ctx);
-      } else
+#else
+        abort();
 #endif
+      } else {
         bitset_dist_current.reset();
+      }
 
       syncSubstrate->set_num_run(run + 1);
       InitializeGraph::go((*hg));
@@ -421,23 +432,23 @@ int main(int argc, char** argv) {
 
   // Verify
   if (verify) {
-#ifdef __GALOIS_HET_CUDA__
     if (personality == CPU) {
-#endif
       for (auto ii = (*hg).masterNodesRange().begin();
            ii != (*hg).masterNodesRange().end(); ++ii) {
         galois::runtime::printOutput("% %\n", (*hg).getGID(*ii),
                                      (*hg).getData(*ii).dist_current);
       }
-#ifdef __GALOIS_HET_CUDA__
     } else if (personality == GPU_CUDA) {
+#ifdef __GALOIS_HET_CUDA__
       for (auto ii = (*hg).masterNodesRange().begin();
            ii != (*hg).masterNodesRange().end(); ++ii) {
         galois::runtime::printOutput("% %\n", (*hg).getGID(*ii),
                                      get_node_dist_current_cuda(cuda_ctx, *ii));
       }
-    }
+#else
+        abort();
 #endif
+    }
   }
 
   return 0;
