@@ -29,6 +29,9 @@
 #ifdef __GALOIS_HET_CUDA__
 #include "cc_push_cuda.h"
 struct CUDA_Context* cuda_ctx;
+#else
+enum { CPU, GPU_CUDA };
+int personality = CPU;
 #endif
 
 constexpr static const char* const REGION_NAME = "ConnectedComp";
@@ -82,16 +85,17 @@ struct InitializeGraph {
   void static go(Graph& _graph) {
     const auto& allNodes = _graph.allNodesRange();
 
-#ifdef __GALOIS_HET_CUDA__
     if (personality == GPU_CUDA) {
+#ifdef __GALOIS_HET_CUDA__
       std::string impl_str("InitializeGraph_" + (syncSubstrate->get_run_identifier()));
       galois::StatTimer StatTimer_cuda(impl_str.c_str(), REGION_NAME);
       StatTimer_cuda.start();
       InitializeGraph_allNodes_cuda(cuda_ctx);
       StatTimer_cuda.stop();
-    } else if (personality == CPU)
+#else
+      abort();
 #endif
-    {
+    } else if (personality == CPU) {
       galois::do_all(galois::iterate(allNodes.begin(), allNodes.end()),
                      InitializeGraph{&_graph}, galois::no_stats(),
                      galois::loopname(
@@ -115,17 +119,18 @@ struct FirstItr_ConnectedComp {
   void static go(Graph& _graph) {
     const auto& nodesWithEdges = _graph.allNodesWithEdgesRange();
     syncSubstrate->set_num_round(0);
-#ifdef __GALOIS_HET_CUDA__
     if (personality == GPU_CUDA) {
+#ifdef __GALOIS_HET_CUDA__
       std::string impl_str("ConnectedComp_" + (syncSubstrate->get_run_identifier()));
       galois::StatTimer StatTimer_cuda(impl_str.c_str(), REGION_NAME);
       StatTimer_cuda.start();
       StatTimer_cuda.stop();
       FirstItr_ConnectedComp_nodesWithEdges_cuda(cuda_ctx);
       StatTimer_cuda.stop();
-    } else if (personality == CPU)
+#else
+      abort();
 #endif
-    {
+    } else if (personality == CPU) {
       galois::do_all(
           galois::iterate(nodesWithEdges), FirstItr_ConnectedComp{&_graph},
           galois::steal(), galois::no_stats(),
@@ -180,8 +185,8 @@ struct ConnectedComp {
     do {
       syncSubstrate->set_num_round(_num_iterations);
       dga.reset();
-#ifdef __GALOIS_HET_CUDA__
       if (personality == GPU_CUDA) {
+#ifdef __GALOIS_HET_CUDA__
         std::string impl_str("ConnectedComp_" + (syncSubstrate->get_run_identifier()));
         galois::StatTimer StatTimer_cuda(impl_str.c_str(), REGION_NAME);
         StatTimer_cuda.start();
@@ -189,9 +194,10 @@ struct ConnectedComp {
         ConnectedComp_nodesWithEdges_cuda(__retval, cuda_ctx);
         dga += __retval;
         StatTimer_cuda.stop();
-      } else if (personality == CPU)
+#else
+        abort();
 #endif
-      {
+      } else if (personality == CPU) {
         galois::do_all(galois::iterate(nodesWithEdges),
                        ConnectedComp(&_graph, dga), galois::no_stats(),
                        galois::steal(),
@@ -251,17 +257,20 @@ struct ConnectedCompSanityCheck {
   void static go(Graph& _graph, galois::DGAccumulator<uint64_t>& dga) {
     dga.reset();
 
-#ifdef __GALOIS_HET_CUDA__
     if (personality == GPU_CUDA) {
+#ifdef __GALOIS_HET_CUDA__
       uint64_t sum;
       ConnectedCompSanityCheck_masterNodes_cuda(sum, cuda_ctx);
       dga += sum;
-    } else
+#else
+      abort();
 #endif
+    } else {
       galois::do_all(galois::iterate(_graph.masterNodesRange().begin(),
                                      _graph.masterNodesRange().end()),
                      ConnectedCompSanityCheck(&_graph, dga), galois::no_stats(),
                      galois::loopname("ConnectedCompSanityCheck"));
+    }
 
     uint64_t num_components = dga.reduce();
 
@@ -339,12 +348,15 @@ int main(int argc, char** argv) {
     ConnectedCompSanityCheck::go(*hg, active_vertices64);
 
     if ((run + 1) != numRuns) {
-#ifdef __GALOIS_HET_CUDA__
       if (personality == GPU_CUDA) {
+#ifdef __GALOIS_HET_CUDA__
         bitset_comp_current_reset_cuda(cuda_ctx);
-      } else
+#else
+        abort();
 #endif
+      } else {
         bitset_comp_current.reset();
+      }
 
       (*syncSubstrate).set_num_run(run + 1);
       InitializeGraph::go((*hg));
@@ -356,23 +368,23 @@ int main(int argc, char** argv) {
 
   // Verify
   if (verify) {
-#ifdef __GALOIS_HET_CUDA__
     if (personality == CPU) {
-#endif
       for (auto ii = (*hg).masterNodesRange().begin();
            ii != (*hg).masterNodesRange().end(); ++ii) {
         galois::runtime::printOutput("% %\n", (*hg).getGID(*ii),
                                      (*hg).getData(*ii).comp_current);
       }
-#ifdef __GALOIS_HET_CUDA__
     } else if (personality == GPU_CUDA) {
+#ifdef __GALOIS_HET_CUDA__
       for (auto ii = (*hg).masterNodesRange().begin();
            ii != (*hg).masterNodesRange().end(); ++ii) {
         galois::runtime::printOutput("% %\n", (*hg).getGID(*ii),
                                      get_node_comp_current_cuda(cuda_ctx, *ii));
       }
-    }
+#else
+      abort();
 #endif
+    }
   }
 
   return 0;
