@@ -29,6 +29,9 @@
 #ifdef __GALOIS_HET_CUDA__
 #include "sssp_pull_cuda.h"
 struct CUDA_Context* cuda_ctx;
+#else
+enum { CPU, GPU_CUDA };
+int personality = CPU;
 #endif
 
 constexpr static const char* const REGION_NAME = "SSSP";
@@ -89,16 +92,17 @@ struct InitializeGraph {
 
   void static go(Graph& _graph) {
     const auto& allNodes = _graph.allNodesRange();
-#ifdef __GALOIS_HET_CUDA__
     if (personality == GPU_CUDA) {
+#ifdef __GALOIS_HET_CUDA__
       std::string impl_str("InitializeGraph_" + (syncSubstrate->get_run_identifier()));
       galois::StatTimer StatTimer_cuda(impl_str.c_str(), REGION_NAME);
       StatTimer_cuda.start();
       InitializeGraph_allNodes_cuda(infinity, src_node, cuda_ctx);
       StatTimer_cuda.stop();
-    } else if (personality == CPU)
+#else
+        abort();
 #endif
-    {
+    } else if (personality == CPU) {
       galois::do_all(galois::iterate(allNodes.begin(), allNodes.end()),
                      InitializeGraph{src_node, infinity, &_graph},
                      galois::no_stats(),
@@ -133,8 +137,8 @@ struct SSSP {
     do {
       syncSubstrate->set_num_round(_num_iterations);
       dga.reset();
-#ifdef __GALOIS_HET_CUDA__
       if (personality == GPU_CUDA) {
+#ifdef __GALOIS_HET_CUDA__
         std::string impl_str("SSSP_" + (syncSubstrate->get_run_identifier()));
         galois::StatTimer StatTimer_cuda(impl_str.c_str(), REGION_NAME);
         StatTimer_cuda.start();
@@ -142,9 +146,10 @@ struct SSSP {
         SSSP_nodesWithEdges_cuda(__retval, cuda_ctx);
         dga += __retval;
         StatTimer_cuda.stop();
-      } else if (personality == CPU)
+#else
+        abort();
 #endif
-      {
+      } else if (personality == CPU) {
         galois::do_all(
             galois::iterate(nodesWithEdges), SSSP{&_graph, dga},
             galois::no_stats(), galois::steal(),
@@ -213,17 +218,19 @@ struct SSSPSanityCheck {
     dgm.reset();
     dgag.reset();
 
-#ifdef __GALOIS_HET_CUDA__
     if (personality == GPU_CUDA) {
-      uint64_t sum, avg;
+#ifdef __GALOIS_HET_CUDA__
+      uint64_t sum;
+      uint64_t avg;
       uint32_t max;
       SSSPSanityCheck_masterNodes_cuda(sum, avg, max, infinity, cuda_ctx);
       dgas += sum;
       dgm.update(max);
       dgag += avg;
-    } else
+#else
+        abort();
 #endif
-    {
+    } else {
       galois::do_all(galois::iterate(_graph.masterNodesRange().begin(),
                                      _graph.masterNodesRange().end()),
                      SSSPSanityCheck(infinity, &_graph, dgas, dgm, dgag),
@@ -318,12 +325,14 @@ int main(int argc, char** argv) {
     SSSPSanityCheck::go(*hg, DGAccumulator_sum, m, dg_avge);
 
     if ((run + 1) != numRuns) {
-#ifdef __GALOIS_HET_CUDA__
       if (personality == GPU_CUDA) {
+#ifdef __GALOIS_HET_CUDA__
         bitset_dist_current_reset_cuda(cuda_ctx);
-      } else
+#else
+        abort();
 #endif
-        bitset_dist_current.reset();
+      } else { bitset_dist_current.reset();
+      }
 
       (*syncSubstrate).set_num_run(run + 1);
       InitializeGraph::go((*hg));
@@ -335,23 +344,23 @@ int main(int argc, char** argv) {
 
   // Verify
   if (verify) {
-#ifdef __GALOIS_HET_CUDA__
     if (personality == CPU) {
-#endif
       for (auto ii = (*hg).masterNodesRange().begin();
            ii != (*hg).masterNodesRange().end(); ++ii) {
         galois::runtime::printOutput("% %\n", (*hg).getGID(*ii),
                                      (*hg).getData(*ii).dist_current);
       }
-#ifdef __GALOIS_HET_CUDA__
     } else if (personality == GPU_CUDA) {
+#ifdef __GALOIS_HET_CUDA__
       for (auto ii = (*hg).masterNodesRange().begin();
            ii != (*hg).masterNodesRange().end(); ++ii) {
         galois::runtime::printOutput("% %\n", (*hg).getGID(*ii),
                                      get_node_dist_current_cuda(cuda_ctx, *ii));
       }
-    }
+#else
+        abort();
 #endif
+    }
   }
 
   return 0;
